+++ src/edu4.WebClient/src/comps/nav/Avatar.js
