+++ src/edu4.WebClient/src/comps/nav/LogoIcon.js
